third_party/WebKit/LayoutTests/http/tests/security/cross-origin-session-storage-allowed.html

<html>
<head>
<script>
if (window.testRunner) {
	testRunner.dumpAsText();
	testRunner.dumpChildFrames();
}
</script>
</head>
<body>
<p>This iframe should not return any errors:</p>
<iframe src="http://localhost:8000/security/resources/cross-origin-iframe-for-session-storage.html"></iframe>
<p>This iframe should not return any errors:</p>
<iframe src="http://127.0.0.1:8000/security/resources/cross-origin-iframe-for-session-storage.html"></iframe>
</body>
</html>
