third_party/blink/web_tests/http/tests/security/cross-origin-session-storage-allowed-expected.txt

This iframe should not return any errors:

This iframe should not return any errors:

--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
No exception

--------
Frame: '<!--framePath //<!--frame1-->-->'
--------
No exception
